Parser: Split out Default Parser Configuration
